Rename dnsiface.Service for how-to-go-fast
