+++ internal/dnsx/dnsiface/dns.go
@@ -1,4 +1,4 @@
-package iface
+package dnsiface
 
 import (
 	"context"
@@ -6,23 +6,23 @@ import (
 	"google.golang.org/api/dns/v1"
 )
 
-// DNS interface used by the dnsx logic.
-type DNS interface {
+// Service interface used by the dnsx logic.
+type Service interface {
 	ResourceRecordSetsGet(ctx context.Context, project string, zone string, name string, type_ string) (*dns.ResourceRecordSet, error)
 	ChangeCreate(ctx context.Context, project string, zone string, change *dns.Change) (*dns.Change, error)
 }
 
-// DNSImpl implements the DNS interface.
-type DNSImpl struct {
+// CloudDNSService implements the DNS Service interface.
+type CloudDNSService struct {
 	Service *dns.Service
 }
 
 // ResourceRecordSetsGet gets an existing resource record set, if present.
-func (d *DNSImpl) ResourceRecordSetsGet(ctx context.Context, project string, zone string, name string, rtype string) (*dns.ResourceRecordSet, error) {
-	return d.Service.ResourceRecordSets.Get(project, zone, name, rtype).Context(ctx).Do()
+func (c *CloudDNSService) ResourceRecordSetsGet(ctx context.Context, project string, zone string, name string, rtype string) (*dns.ResourceRecordSet, error) {
+	return c.Service.ResourceRecordSets.Get(project, zone, name, rtype).Context(ctx).Do()
 }
 
 // ChangeCreate applies the given change set.
-func (d *DNSImpl) ChangeCreate(ctx context.Context, project string, zone string, change *dns.Change) (*dns.Change, error) {
-	return d.Service.Changes.Create(project, zone, change).Context(ctx).Do()
+func (c *CloudDNSService) ChangeCreate(ctx context.Context, project string, zone string, change *dns.Change) (*dns.Change, error) {
+	return c.Service.Changes.Create(project, zone, change).Context(ctx).Do()
 }
